rename adapter dir
