+++ packages/reports/src/reports/blocks/types/adapters/Generator.js
