+++ packages/reports/src/reports/blocks/types/adapters/Selector.js
